backend/requirements.txt
crewai
crewai_tools
langchain_ollama
fastapi
uvicorn[standard]
python-dotenv
chromadb
sentence-transformers